+++ CHANGELOG.md
@@ -33,6 +33,7 @@
     so it's possible to override `password_length` at runtime. (@manojmj92)
 * bug fixes
   * Make `Devise` work without `ActionMailer` when `Zeitwerk` autoloader is used.
+  * Set `attempted_path` in `warden.options` before calling failure app in controller test helpers.
 
 Please check [4-stable](https://github.com/heartcombo/devise/blob/4-stable/CHANGELOG.md)
 for previous changes.